# api-docs.yaml
openapi: "3.0.3"
info:
  title: "CRM System API"
  version: "1.0.0"
  description: "Sistema de gerenciamento de relacionamento com clientes - API completa para operações CRUD e relatórios de dashboard"
  contact:
    name: "CRM System Support"
    email: "support@crm-system.com"
  license:
    name: "ISC"
    url: "https://opensource.org/licenses/ISC"
servers:
  - url: "http://localhost:3000/api"
    description: "Development server"
  - url: "https://api.crm-system.com/api"
    description: "Production server"
components:
  schemas:
...
    MonthlyRevenueResponse:
      type: "object"
      additionalProperties:
        type: "number"
        minimum: 0
      description: "Monthly revenue data response"
      example:
        Janeiro: 45000
        Fevereiro: 52000
        Março: 48000
    MoreSalesByResponsibleResponse:
      type: "array"
      items:
        type: "object"
        properties:
          responsibleId:
            type: "string"
            minLength: 1
          responsibleName:
            type: "string"
            minLength: 1
          saleValue:
            type: "number"
            minimum: 0
          closedDealsCount:
            type: "integer"
            minimum: 0
        required:
          - responsibleId
          - responsibleName
          - saleValue
          - closedDealsCount
      description: "Sales data grouped by responsible person"
      example:
        - responsibleId: "123e4567-e89b-12d3-a456-426614174000"
          responsibleName: "John Doe"
          saleValue: 125000
          closedDealsCount: 8
        - responsibleId: "456e7890-e89b-12d3-a456-426614174000"
          responsibleName: "Jane Smith"
          saleValue: 98000
          closedDealsCount: 6
    SalesFunnelResponse:
      type: "object"
      additionalProperties:
        type: "integer"
        minimum: 0
      description: "Sales funnel data by stage"
      example:
        Prospecting: 25
        Qualification: 18
        Proposal: 12
        Negotiation: 8
        Closed Won: 15
        Closed Lost: 5
    TotalRevenueResponse:
      type: "object"
      properties:
        total:
          type: "number"
          minimum: 0
      required:
        - total
      description: "Total revenue response"
      example:
        total: 245000
    ActiveAccountsResponse:
      type: "object"
      properties:
        total:
          type: "integer"
          minimum: 0
      required:
        - total
      description: "Active accounts count response"
      example:
        total: 156
    NewBusinessResponse:
      type: "object"
      properties:
        count:
          type: "integer"
          minimum: 0
      required:
        - count
      description: "New business count response"
      example:
        count: 23
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "JWT token for authentication. Include the token in the Authorization header as: Bearer <token>"
  responses:
    UnauthorizedError:
      description: "Authentication information is missing or invalid"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Unauthorized"
              status:
                type: "integer"
                example: 401
...
security:
  - bearerAuth:
...
paths:
...
  /api/dashboard/revenue-per-period:
    get:
      operationId: "getRevenuePerPeriod"
      summary: "Get monthly revenue totals for a specific period"
      description: "Retrieves monthly revenue data aggregated by month for the specified period"
      tags:
        - Dashboard
      parameters:
        - name: "period"
          in: "query"
          required: false
          schema:
            type: "string"
            enum:
              - THIS_MONTH
              - THIS_YEAR
              - LAST_QUARTER
            default: "THIS_YEAR"
          description: "Period for revenue data (optional, defaults to THIS_YEAR)"
          example: "THIS_YEAR"
      security:
        - bearerAuth:
      responses:
        200:
          description: "Monthly revenue data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyRevenueResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
  /api/dashboard/more-sales-by-responsible:
    get:
      operationId: "getMoreSalesByResponsible"
      summary: "Get sales performance by responsible users"
      description: "Retrieves sales performance data grouped by responsible users for the specified period"
      tags:
        - Dashboard
      parameters:
        - name: "period"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - THIS_MONTH
              - THIS_YEAR
              - LAST_QUARTER
          description: "Time period for sales data"
          example: "THIS_MONTH"
      security:
        - bearerAuth:
      responses:
        200:
          description: "Sales performance data by responsible"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoreSalesByResponsibleResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
  /api/dashboard/sales-funnel:
    get:
      operationId: "getSalesFunnel"
      summary: "Get sales funnel distribution by stage"
      description: "Retrieves sales funnel data showing distribution of opportunities by stage"
      tags:
        - Dashboard
      parameters:
        - name: "period"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - THIS_MONTH
              - THIS_YEAR
              - LAST_QUARTER
          description: "Time period for funnel data"
          example: "THIS_MONTH"
      security:
        - bearerAuth:
      responses:
        200:
          description: "Sales funnel data by stage"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesFunnelResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
  /api/dashboard/total-revenue:
    get:
      operationId: "getTotalRevenue"
      summary: "Get total revenue from closed-won deals"
      description: "Retrieves total revenue from all closed-won business deals for the specified period"
      tags:
        - Dashboard
      parameters:
        - name: "period"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - THIS_MONTH
              - THIS_YEAR
              - LAST_QUARTER
          description: "Time period for revenue calculation"
          example: "THIS_MONTH"
      security:
        - bearerAuth:
      responses:
        200:
          description: "Total revenue data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotalRevenueResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
  /api/dashboard/active-accounts:
    get:
      operationId: "getActiveAccounts"
      summary: "Get total count of active accounts"
      description: "Retrieves the total number of accounts with active status"
      tags:
        - Dashboard
      parameters:
      security:
        - bearerAuth:
      responses:
        200:
          description: "Active accounts count"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveAccountsResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
  /api/dashboard/new-business:
    get:
      operationId: "getNewBusiness"
      summary: "Get count of new businesses created"
      description: "Retrieves the count of new business opportunities created in the specified period"
      tags:
        - Dashboard
      parameters:
        - name: "period"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - THIS_MONTH
              - THIS_YEAR
              - LAST_QUARTER
          description: "Time period for new business count"
          example: "THIS_MONTH"
      security:
        - bearerAuth:
      responses:
        200:
          description: "New business count"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewBusinessResponse"
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/UnauthorizedError"
...